Updated to 0.3.0 Updated to es6 and remove support for instances prior to API
