+++ assets/javascripts/lib/discourse-markdown/checklist.js
@@ -1,0 +1,27 @@
+import { registerOption } from 'pretty-text/pretty-text';
+
+registerOption((siteSettings, opts) => {
+  opts.features['checkist'] = !!siteSettings.checklist_enabled;
+});
+
+function replaceChecklist(text) {
+  text = text || "";
+  text = text.replace(/\[\s?\]/ig, '<span class="chcklst-box fa fa-square-o"></span>');
+  text = text.replace(/\[_\]/ig, '<span class="chcklst-box fa fa-square"></span>');
+  text = text.replace(/\[-\]/ig, '<span class="chcklst-box fa fa-minus-square-o"></span>');
+  text = text.replace(/\[x\]/ig, '<span class="chcklst-box checked fa fa-check-square"></span>');
+  text = text.replace(/\[\*\]/ig, '<span class="chcklst-box checked fa fa-check-square-o"></span>');
+  text = text.replace(/!<span class="chcklst-box (checked )?(fa fa-(square-o|square|minus-square-o|check-square|check-square-o))"><\/span>\(/ig, "![](");
+  return text;
+}
+
+export function setup(helper) {
+  helper.inlineBetween({
+    between: "--",
+    emitter: function(contents) {
+      return ["span", {"class": "chcklst-stroked"}].concat(contents);
+    }
+  });
+  helper.whiteList([ 'span.chcklst-stroked', 'span.chcklst-*' ]);
+  helper.addPreProcessor(replaceChecklist);
+}
